🏗️ structure: Refactor component folder structure based on functionality grouping #56
