+++ src/challenge/components/action/StartButton.tsx
