+++ src/challenge/components/dataDisplay/AppliedMemberAvatars.tsx
