+++ src/challenge/components/dataDisplay/ChallengeCard.tsx
