+++ src/challenge/components/dataDisplay/RestMemberSlotChip.tsx
