+++ src/challenge/components/dataDisplay/Suggestion.tsx
