+++ src/challenge/components/status/Error.tsx
