+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import { Banner } from '@banner/components/Banner';
-import { ChallengeList } from '@challenge/components/ChallengeList';
-import { ChallengeListError } from '@challenge/components/Error';
+import { ChallengeList } from '@challenge/components/dataDisplay/ChallengeGridList';
+import { ChallengeListError } from '@challenge/components/status/Error';
 import { mainChallengeButton } from '@challenge/styles/challengeStyle';
 import { Layout } from '@shared/components/dataDisplay/FlexLayout';
 import { CompositionBoundaryReactQuery } from '@shared/boundaries';
